perf(nc-gui): eliminate form-builder field-switch lag
